refactor: refactor client read stream class implementation
